Added some basic window management functions
Mash + C -> Center the frontmost window
Mash + M -> Maximize the frontmost window (toggle!)

When maximizing a window, the original frame is stored in memory. When
pressing the hotkey a second time, the window is reset to the original
frame. Simple, but effective.

The code is kept simple on purpose. If you manually resize a maximized
window, it will not be maximized. Instead it will be reset to the
previous dimensions. No problem for me.