Make application activation more robust.
Sometimes activating an application doesn't do anything. This is typically the
case for OmniFocus, which has a global popup window. If that window was the last
one shown, activation doesn't do anything.

Fallback in case no window was shown: try to select the 'Window -> Main Window'
menu item. At least OmniFocus has it...